Tweak requirements & description for flint hand drill
It's essentially a fire bow drill, but with a flint blade, so make the wood requirement include the spindle & plate.
Include cutting for shaping the wood, the cordage is both for the bow drill and the binding, secured with bone glue.

This makes it not a 10 minute craft item, but not a multiple-season gated item due to forging requirements.